add support for ecs platform_type
